feat(frontend): created an UV project 		Added a prototype main page and two extra pages each  simulating a room 		Added dummy Data 		Improving Code by capsulating certain functions 		corrected optimal_room_conditions.md to use VOC-index 		instead of PPM 		added Vox index in parameter.json
