+++ frontend/example_room_data.py
@@ -1,0 +1,13 @@
+sensor_data_room_1 = {
+    "temperature_inside": 20,   
+    "humidity_inside": 35,      
+    "voc_index": 120,        
+    "noise_level": 55           
+}
+
+sensor_data_room_2 = {
+    "temperature_inside": 30,   
+    "humidity_inside": 30,      
+    "voc_index": 220,           
+    "noise_level": 60          
+}
