+++ frontend/pages/room1.py
@@ -1,0 +1,6 @@
+
+from functions import define_page
+from example_room_data import sensor_data_room_1
+
+
+define_page(1, sensor_data_room_1)
