+++ frontend/pages/room2.py
@@ -1,0 +1,7 @@
+from functions import define_page
+from example_room_data import sensor_data_room_2
+
+
+
+
+define_page(2, sensor_data_room_2)
